+++ components/_util/__tests__/util.test.js
@@ -12,7 +12,6 @@ import {
   throttleByAnimationFrameDecorator,
 } from '../throttleByAnimationFrame';
 import TransButton from '../transButton';
-import Wave from '../wave';
 
 describe('Test utils function', () => {
   describe('throttle', () => {
@@ -137,60 +136,6 @@ describe('Test utils function', () => {
           });
         });
       });
-    });
-  });
-
-  describe('wave', () => {
-    it('bindAnimationEvent should return when node is null', () => {
-      const wrapper = mount(
-        <Wave>
-          <button type="button" disabled>
-            button
-          </button>
-        </Wave>,
-      )
-        .find(Wave)
-        .instance();
-      expect(wrapper.bindAnimationEvent()).toBe(undefined);
-    });
-
-    it('bindAnimationEvent.onClick should return when children is hidden', () => {
-      const wrapper = mount(
-        <Wave>
-          <button type="button" style={{ display: 'none' }}>
-            button
-          </button>
-        </Wave>,
-      )
-        .find(Wave)
-        .instance();
-      expect(wrapper.bindAnimationEvent()).toBe(undefined);
-    });
-
-    it('bindAnimationEvent.onClick should return when children is input', () => {
-      const wrapper = mount(
-        <Wave>
-          <input />
-        </Wave>,
-      )
-        .find(Wave)
-        .instance();
-      expect(wrapper.bindAnimationEvent()).toBe(undefined);
-    });
-
-    it('should not throw when click it', () => {
-      expect(() => {
-        const wrapper = mount(
-          <Wave>
-            <div />
-          </Wave>,
-        );
-        wrapper.simulate('click');
-      }).not.toThrow();
-    });
-
-    it('should not throw when no children', () => {
-      expect(() => mount(<Wave />)).not.toThrow();
     });
   });
