+++ components/_util/__tests__/wave.test.js
@@ -1,7 +1,7 @@
 import { mount } from 'enzyme';
 import React from 'react';
 import mountTest from '../../../tests/shared/mountTest';
-import { sleep } from '../../../tests/utils';
+import { render, sleep } from '../../../tests/utils';
 import ConfigProvider from '../../config-provider';
 import Wave from '../wave';
 
@@ -147,5 +147,54 @@ describe('Wave component', () => {
     const styles = wrapper.find('button').getDOMNode().getRootNode().getElementsByTagName('style');
     expect(styles[0].getAttribute('nonce')).toBe('YourNonceCode');
     wrapper.unmount();
+  });
+
+  it('bindAnimationEvent should return when node is null', () => {
+    const ref = React.createRef();
+    render(
+      <Wave ref={ref}>
+        <button type="button" disabled>
+          button
+        </button>
+      </Wave>,
+    );
+    expect(ref.current?.bindAnimationEvent()).toBe(undefined);
+  });
+
+  it('bindAnimationEvent.onClick should return when children is hidden', () => {
+    const ref = React.createRef();
+    render(
+      <Wave ref={ref}>
+        <button type="button" style={{ display: 'none' }}>
+          button
+        </button>
+      </Wave>,
+    );
+    expect(ref.current?.bindAnimationEvent()).toBe(undefined);
+  });
+
+  it('bindAnimationEvent.onClick should return when children is input', () => {
+    const ref = React.createRef();
+    render(
+      <Wave ref={ref}>
+        <input />
+      </Wave>,
+    );
+    expect(ref.current?.bindAnimationEvent()).toBe(undefined);
+  });
+
+  it('should not throw when click it', () => {
+    expect(() => {
+      const wrapper = mount(
+        <Wave>
+          <div />
+        </Wave>,
+      );
+      wrapper.simulate('click');
+    }).not.toThrow();
+  });
+
+  it('should not throw when no children', () => {
+    expect(() => mount(<Wave />)).not.toThrow();
   });
 });
